requirements.txt
anyio==4.9.0
APScheduler==3.10.4
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
cryptography==45.0.3
firebase-admin==6.9.0
google-api-core==2.25.0
google-api-python-client==2.171.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.72.1
grpcio-status==1.72.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.26.0
hyperframe==6.1.0
idna==3.10
instagrapi==2.0.0
msgpack==1.1.0
proto-plus==1.26.1
protobuf==6.31.1
pyaes==1.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodomex==3.18.0
pydantic==1.10.9
PyJWT==2.10.1
pyparsing==3.2.3
Pyrogram==2.0.106
PySocks==1.7.1
python-telegram-bot==20.8
pytz==2025.2
requests==2.32.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
typing_extensions==4.14.0
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.4.0
